refactor(frontend): apply review conventions and 14px expert details
