+++ autogpt_platform/frontend/src/app/(platform)/team/[expertId]/components/ExpertWorkflowActions.tsx
@@ -1,23 +1,15 @@
 "use client";
 
 import { ExpertWorkflowRef } from "@/app/api/__generated__/models/expertWorkflowRef";
-import { GraphExecutionMeta } from "@/app/api/__generated__/models/graphExecutionMeta";
-import { LibraryAgent } from "@/app/api/__generated__/models/libraryAgent";
-import { RunAgentModal } from "@/app/(platform)/library/agents/[id]/components/NewAgentLibraryView/components/modals/RunAgentModal/RunAgentModal";
 import { Button } from "@/components/atoms/Button/Button";
 import { Icon } from "@/components/atoms/Icon/Icon";
-import {
-  PencilEdit02Icon,
-  PlayIcon,
-  SparklesIcon,
-} from "@hugeicons/core-free-icons";
+import { PencilEdit02Icon, SparklesIcon } from "@hugeicons/core-free-icons";
 import { ExpertWorkflowCardMenu } from "./ExpertWorkflowCardMenu";
-import { ACTION_BUTTON_CLASS } from "@/app/(platform)/team/helpers";
 
 const ICON_BUTTON_CLASS =
   "h-7 w-7 rounded-md border-transparent bg-white/90 p-0 text-zinc-700 hover:border-transparent hover:bg-white";
 
-interface ActionsProps {
+interface Props {
   workflow: ExpertWorkflowRef;
   expertId: string;
   name: string;
@@ -35,7 +27,7 @@ export function ExpertWorkflowActions({
   chatHref,
   buttonClassName = ICON_BUTTON_CLASS,
   menuClassName = buttonClassName,
-}: ActionsProps) {
+}: Props) {
   return (
     <>
       {builderHref ? (
@@ -67,40 +59,5 @@ export function ExpertWorkflowActions({
         triggerClassName={menuClassName}
       />
     </>
-  );
-}
-
-interface RunProps {
-  agent: LibraryAgent;
-  isTriggerWorkflow: boolean;
-  variant?: "primary" | "secondary";
-  onRunCreated: (execution: GraphExecutionMeta) => void;
-  onTriggerSetup: () => void;
-}
-
-export function ExpertWorkflowRunButton({
-  agent,
-  isTriggerWorkflow,
-  onRunCreated,
-  onTriggerSetup,
-  variant = "secondary",
-}: RunProps) {
-  return (
-    <RunAgentModal
-      agent={agent}
-      onRunCreated={onRunCreated}
-      onTriggerSetup={onTriggerSetup}
-      triggerSlot={
-        <Button
-          type="button"
-          variant={variant}
-          size="small"
-          className={ACTION_BUTTON_CLASS}
-          leftIcon={<Icon icon={PlayIcon} size={14} />}
-        >
-          {isTriggerWorkflow ? "Set up trigger" : "Run"}
-        </Button>
-      }
-    />
   );
 }
