+++ autogpt_platform/frontend/src/app/(platform)/team/[expertId]/components/ExpertWorkflowListItem.tsx
@@ -7,10 +7,8 @@ import { safeHumanizeCronExpression } from "@/lib/cron-expression-utils";
 import { cn } from "@/lib/utils";
 import { Activity01Icon } from "@hugeicons/core-free-icons";
 import NextLink from "next/link";
-import {
-  ExpertWorkflowActions,
-  ExpertWorkflowRunButton,
-} from "./ExpertWorkflowActions";
+import { ExpertWorkflowActions } from "./ExpertWorkflowActions";
+import { ExpertWorkflowRunButton } from "./ExpertWorkflowRunButton";
 import { useExpertWorkflowCard } from "./useExpertWorkflowCard";
 import { WorkflowCredentialStack } from "./WorkflowCredentialStack";
